# Agentic-Sentry/openapi.yaml
openapi: "3.0.3"
info:
  title: User Service
  description: >
    Modernised user-lookup microservice.
    Replaces the legacy Flask monolith with a FastAPI service that uses
    parameterised queries, Pydantic validation, and environment-driven secrets.
  version: "2.0.0"
  contact:
    name: Platform Engineering

servers:
  - url: http://localhost:8000
    description: Local development server

tags:
  - name: users
    description: User management operations
  - name: ops
    description: Operational / observability endpoints

components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key

  schemas:
    UserResponse:
      type: object
      required: [id, username, email]
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: alice
        email:
          type: string
          format: email
          example: alice@example.com

    HealthResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          example: ok

    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          example: User with id=42 not found

    ValidationErrorResponse:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
              msg:
                type: string
              type:
                type: string

security:
  - ApiKeyHeader: []

paths:
  /health:
    get:
      summary: Liveness probe
      description: Returns 200 OK when the service is running. No authentication required.
      operationId: healthCheck
      tags: [ops]
      security: []          # public endpoint — no API key needed
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /api/v2/users/{user_id}:
    get:
      summary: Get user by ID
      description: Retrieve a single user record by their integer primary-key ID.
      operationId: getUser
      tags: [users]
      security:
        - ApiKeyHeader: []
      parameters:
        - name: user_id
          in: path
          required: true
          description: Integer primary-key identifier of the user.
          schema:
            type: integer
            example: 1
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error — e.g. non-integer user_id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"